refactor(humanize): move to ./internal and better docs
