+++ internal/cmd/miniooni/libminiooni.go
@@ -17,10 +17,10 @@ import (
 
 	"github.com/apex/log"
 	"github.com/ooni/probe-cli/v3/internal/engine"
-	"github.com/ooni/probe-cli/v3/internal/engine/humanizex"
 	"github.com/ooni/probe-cli/v3/internal/engine/legacy/assetsdir"
 	"github.com/ooni/probe-cli/v3/internal/engine/model"
 	"github.com/ooni/probe-cli/v3/internal/engine/netx/selfcensor"
+	"github.com/ooni/probe-cli/v3/internal/humanize"
 	"github.com/ooni/probe-cli/v3/internal/version"
 	"github.com/pborman/getopt/v2"
 )
@@ -377,8 +377,8 @@ func MainWithConfiguration(experimentName string, currentOptions Options) {
 	defer func() {
 		sess.Close()
 		log.Infof("whole session: recv %s, sent %s",
-			humanizex.SI(sess.KibiBytesReceived()*1024, "byte"),
-			humanizex.SI(sess.KibiBytesSent()*1024, "byte"),
+			humanize.SI(sess.KibiBytesReceived()*1024, "byte"),
+			humanize.SI(sess.KibiBytesSent()*1024, "byte"),
 		)
 	}()
 	log.Debugf("miniooni temporary directory: %s", sess.TempDir())
@@ -426,8 +426,8 @@ func MainWithConfiguration(experimentName string, currentOptions Options) {
 	experiment := builder.NewExperiment()
 	defer func() {
 		log.Infof("experiment: recv %s, sent %s",
-			humanizex.SI(experiment.KibiBytesReceived()*1024, "byte"),
-			humanizex.SI(experiment.KibiBytesSent()*1024, "byte"),
+			humanize.SI(experiment.KibiBytesReceived()*1024, "byte"),
+			humanize.SI(experiment.KibiBytesSent()*1024, "byte"),
 		)
 	}()
